rename serverUrl context to server and include displayName
