+++ app/components/post_list/post/body/content/embedded_bindings/button_binding/index.tsx
@@ -10,7 +10,7 @@ import {map} from 'rxjs/operators';
 import {doAppCall, postEphemeralCallResponseForPost} from '@actions/remote/apps';
 import {AppExpandLevels, AppBindingLocations, AppCallTypes, AppCallResponseTypes} from '@constants/apps';
 import {MM_TABLES, SYSTEM_IDENTIFIERS} from '@constants/database';
-import {useServerUrl} from '@context/server_url';
+import {useServerUrl} from '@context/server';
 import {createCallContext, createCallRequest} from '@utils/apps';
 import {getStatusColors} from '@utils/message_attachment_colors';
 import {preventDoubleTap} from '@utils/tap';
