+++ app/components/post_list/post/body/content/message_attachments/action_menu/index.tsx
@@ -5,7 +5,7 @@ import React, {useCallback, useState} from 'react';
 
 import {selectAttachmentMenuAction} from '@actions/local/post';
 import AutocompleteSelector from '@components/autocomplete_selector';
-import {useServerUrl} from '@context/server_url';
+import {useServerUrl} from '@context/server';
 
 type Props = {
     dataSource?: string;
