+++ app/components/team_sidebar/add_team/team_list_item/team_list_item.tsx
@@ -8,7 +8,7 @@ import {addUserToTeam} from '@actions/remote/team';
 import TeamIcon from '@components/team_sidebar/team_list/team_item/team_icon';
 import TouchableWithFeedback from '@components/touchable_with_feedback';
 import {Navigation} from '@constants';
-import {useServerUrl} from '@context/server_url';
+import {useServerUrl} from '@context/server';
 import {useTheme} from '@context/theme';
 import {makeStyleSheetFromTheme} from '@utils/theme';
