+++ app/screens/custom_status/index.tsx
@@ -19,7 +19,7 @@ import TabletTitle from '@components/tablet_title';
 import {CustomStatusDuration, Events, Screens} from '@constants';
 import {SET_CUSTOM_STATUS_FAILURE} from '@constants/custom_status';
 import {MM_TABLES, SYSTEM_IDENTIFIERS} from '@constants/database';
-import {withServerUrl} from '@context/server_url';
+import {withServerUrl} from '@context/server';
 import {withTheme} from '@context/theme';
 import {dismissModal, goToScreen, mergeNavigationOptions, showModal} from '@screens/navigation';
 import {getCurrentMomentForTimezone, getRoundedTime, isCustomStatusExpirySupported, safeParseJSON} from '@utils/helpers';
